requirements.txt
aiofiles==0.8.0
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
black==21.12b0
click==8.0.3
colorama==0.4.4
et-xmlfile==1.1.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
tomli==1.2.2
typing_extensions==4.0.1
uvicorn==0.15.0
